State: Move getJetpackSettingsSaveRequestStatus to global selectors directory
