+++ client/state/jetpack/settings/selectors.js
@@ -4,15 +4,9 @@
 import { get } from 'lodash';
 
 /**
- * Returns the status of the last Jetpack site settings save request
- *
- * @param  {Object}  state  Global state tree
- * @param  {Number}  siteId Site ID
- * @return {String}         The request status (peding, success or error)
+ * External dependencies
  */
-export function getJetpackSettingsSaveRequestStatus( state, siteId ) {
-	return get( state.jetpack.settings.saveRequests, [ siteId, 'status' ] );
-}
+import { getJetpackSettingsSaveRequestStatus } from 'state/selectors';
 
 /**
  * Returns true fi the save Jetpack site settings requests is successful
